fix: Fix Sobel Operator and Add Erode

+++ ipp/src/filtering/filters.h
@@ -21,7 +21,7 @@ void GaussianBlur(Image* img, u32* buf, float sigma, size_t r);
 void MeanFilter(Image* img, u32* buf, size_t r);
 void BilateralFilter(Image* img, u32* buf, size_t r, float col, float sp);
 void MedianFilter(Image* img, u32* buf, size_t block);
-void SobelOperator(const Image* img, u32* out, float* dirs, u32* max_mag);
+void SobelOperator(const Image* img, u32* out, float* dirs, float* max_mag);
 
 void GammaFilter(Image* img, float f);
 void Dilate(Image* img, u32* buf, size_t block);
